# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/auth": {
            "patch": {
                "description": "Refresh the access and refresh tokens using a valid refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Reissue access and refresh tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh Token",
                        "name": "RefreshToken",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "토큰 재발급 성공",
                        "schema": {
                            "$ref": "#/definitions/output.TokenOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "description": "email과 password를 통해 로그인 후 토큰 발급.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign In for a user",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.SignInInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "로그인 성공",
                        "schema": {
                            "$ref": "#/definitions/output.TokenOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "description": "email인증 후 학과 정보를 통해 사용자 회원가입.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign Up for a new user",
                "parameters": [
                    {
                        "description": "User information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/input.SignUpInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constant.Authority": {
            "type": "string",
            "enum": [
                "ROLE_STUDENT",
                "ROLE_STUDENT_COUNCIL"
            ],
            "x-enum-varnames": [
                "ROLE_STUDENT",
                "ROLE_STUDENT_COUNCIL"
            ]
        },
        "input.SignInInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "input.SignUpInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "output.TokenOutput": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "accessTokenExp": {
                    "type": "string"
                },
                "authority": {
                    "$ref": "#/definitions/constant.Authority"
                },
                "refreshToken": {
                    "type": "string"
                },
                "refreshTokenExp": {
                    "type": "string"
                }
            }
        }
    }
}